micedcer_api: Add LLDP and stale time functions
PortLand used to detect the layers of the topology by checking the reception of LLDP messages on switches. It is also important to note that most approaches had to handle with ARP stale time in case a host needs to check again for the destination in its ARP cache. This version incorporates both features.